feat: rename notice references to announcement in dashboard tiles
