+++ src/dashboard/tiles/__init__.py
@@ -1,6 +1,6 @@
 from ._actions import actions
-from ._notice import notice
+from ._announcement import announcement
 from ._security import security
 from ._account_info import account_info
 
-__all__ = ["actions", "notice", "security", "account_info"]
+__all__ = ["actions", "announcement", "security", "account_info"]
