+++ src/dashboard/tiles/_announcement.py
